clean code

+++ mindspore/ops/kernel/cpu/scatter_cpu_kernel.h
@@ -67,7 +67,7 @@ class ScatterCpuKernelMod : public NativeCpuKernelMod, public MatchKernelHelper<
   size_t input_dims_{0};
   int64_t axis_{0};
   std::vector<size_t> output_stride_{};
-  std::vector<size_t> indices_stride_{};
+  std::vector<int> indices_stride_{};
   Reduce reduction_type_{Reduce::REDUCE_NONE};
   std::string input_name_{"input"};
   std::string axis_name_{"dim"};
