+++ mindspore/ops/kernel/cpu/scatter_value_cpu_kernel.cc
@@ -97,7 +97,7 @@ int ScatterValueCpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs,
   // calculate indices_stride
   indices_stride_.resize(input_dims_, 1);
   for (size_t i = input_dims_ - 1; i > 0; --i) {
-    indices_stride_[i - 1] = indices_stride_[i] * static_cast<size_t>(indices_shape[i]);
+    indices_stride_[i - 1] = indices_stride_[i] * indices_shape[i];
   }
 
   // calculate output_stride
@@ -112,7 +112,7 @@ template <typename T, typename S, typename ReductionT>
 bool ScatterValueCpuKernelMod::Scatter(const ReductionT &reduction_func, T *output, const S *indices, T src) {
   auto task = [reduction_func, output, indices, src, this](size_t start, size_t end) {
     for (size_t index = start; index < end; index++) {
-      int remain = index;
+      int remain = static_cast<int>(index);
       int output_offset = 0;
       for (size_t i = 0; i < this->input_dims_; ++i) {
         int output_dim_index = remain / this->indices_stride_[i];
@@ -126,7 +126,7 @@ bool ScatterValueCpuKernelMod::Scatter(const ReductionT &reduction_func, T *outp
             output_dim_index += this->input_axis_size_;
           }
         }
-        output_offset += this->output_stride_[i] * output_dim_index;
+        output_offset += static_cast<int>(this->output_stride_[i]) * output_dim_index;
       }
       reduction_func(output + output_offset, src);
     }
